--- a/core/shared/src/main/scala/org/http4s/QueryOps.scala
+++ b/core/shared/src/main/scala/org/http4s/QueryOps.scala
@@ -35,11 +35,14 @@ trait QueryOps {
 
   /** alias for withQueryParam */
   def +?[T: QueryParam]: Self =
-    _withQueryParam(QueryParam[T].key, Nil)
+    _withQueryParam(QueryParam[T].key, None)
 
   /** alias for withQueryParam */
   def +*?[T: QueryParamKeyLike: QueryParamEncoder](value: T): Self =
-    _withQueryParam(QueryParamKeyLike[T].getKey(value), QueryParamEncoder[T].encode(value) :: Nil)
+    _withQueryParam(
+      QueryParamKeyLike[T].getKey(value),
+      Some(QueryParamEncoder[T].encode(value).value),
+    )
 
   /** alias for withQueryParam */
   def +*?[T: QueryParam: QueryParamEncoder](values: collection.Seq[T]): Self =
@@ -51,7 +54,7 @@ trait QueryOps {
 
   /** alias for withQueryParam */
   def +?[K: QueryParamKeyLike](name: K): Self =
-    _withQueryParam(QueryParamKeyLike[K].getKey(name), Nil)
+    _withQueryParam(QueryParamKeyLike[K].getKey(name), None)
 
   /** alias for withQueryParam
     *
@@ -134,14 +137,14 @@ trait QueryOps {
     * replaced with an empty list.
     */
   def withQueryParam[T: QueryParam]: Self =
-    _withQueryParam(QueryParam[T].key, Nil)
+    _withQueryParam(QueryParam[T].key, None)
 
   /** Creates a new `Self` with the specified parameter in the [[Query]].
     * If a parameter with the given `key` already exists the values will be
     * replaced with an empty list.
     */
   def withQueryParam[K: QueryParamKeyLike](key: K): Self =
-    _withQueryParam(QueryParamKeyLike[K].getKey(key), Nil)
+    _withQueryParam(QueryParamKeyLike[K].getKey(key), None)
 
   /** Creates maybe a new `Self` with the specified parameter in the [[Query]].
     * If a parameter with the given `key` already exists the values will be
@@ -149,7 +152,10 @@ trait QueryOps {
     * instance of `Self` will be returned.
     */
   def withQueryParam[T: QueryParamEncoder, K: QueryParamKeyLike](key: K, value: T): Self =
-    _withQueryParam(QueryParamKeyLike[K].getKey(key), QueryParamEncoder[T].encode(value) :: Nil)
+    _withQueryParam(
+      QueryParamKeyLike[K].getKey(key),
+      Some(QueryParamEncoder[T].encode(value).value),
+    )
 
   /** Creates maybe a new `Self` with the specified parameters in the [[Query]].
     * If a parameter with the given `key` already exists the values will be
@@ -183,6 +189,17 @@ trait QueryOps {
   ): Self =
     params.foldLeft(self) { case (s, (k, v)) =>
       replaceQuery(Query.fromVector(s.withQueryParam(k, v).query.toVector))
+    }
+
+  private def _withQueryParam(
+      name: QueryParameterKey,
+      value: Option[String],
+  ): Self =
+    if (query == Query.blank || query == Query.empty) {
+      replaceQuery(Query(name.value -> value))
+    } else {
+      val baseQuery = query.toVector.filter(_._1 != name.value)
+      replaceQuery(Query.fromVector(baseQuery :+ (name.value -> value)))
     }
 
   private def _withQueryParam(