@danicheg 2022-12-18
Only now I have realized that we could do it in 0.23. So, if we accept this, I will do a backport.